+++ packages/vite-app-ts/src/components/common/faucet-hint-button.tsx
@@ -9,8 +9,8 @@ import { utils } from 'ethers';
 import React, { FC, useContext, useMemo, useState } from 'react';
 import { useDebounce } from 'use-debounce';
 
-import { IScaffoldAppProviders } from '~~/components/main/hooks/useScaffoldAppProviders';
-import { FAUCET_ENABLED } from '~~/config/appConfig';
+import { IScaffoldAppProviders } from '~~/hooks/use-scaffold-app-providers';
+import { FAUCET_ENABLED } from '~~/config/app-config';
 
 /**
  * Is Faucet available?
