+++ packages/vite-app-ts/src/config/app-config.ts
