+++ packages/vite-app-ts/src/config/contract-connector-config.ts
