+++ packages/vite-app-ts/src/config/external-contracts-config.ts
