+++ packages/vite-app-ts/src/functions/get-network-info.ts
