+++ packages/vite-app-ts/src/functions/index.ts
@@ -1,2 +1,2 @@
-export * from './contractsByNetworkName';
-export * from './getNetworkInfo';
+export * from './contracts-by-network-name';
+export * from './get-network-info';
